Include Scry for Oracle Infrastructure
Am creating a PR to request Scry to be shown in the available tools for devs to use for oracles. The platform is the ONLY option for developers to deploy their own oracles with custom signer sets, nodes and fully self host, with no reliance on some token, permissioned creation for feeds or even code needed. Allows setup in <15m and allows for 100s of feeds deployed with almost no time needed.

Fully open source.